+++ Tasks/Hatsak/Task3/Task3_6_tests.py
